+++ crates/aranya-aqc-util/src/testing.rs
@@ -18,7 +18,7 @@ use core::{
 
 use aranya_crypto::{
     self,
-    aqc::{BidiChannelId, UniChannelId},
+    aqc::{BidiPskId, CipherSuiteId, UniPskId},
     csprng::Random,
     engine::WrappedKey,
     keystore::{memstore, Entry, Occupied, Vacant},
@@ -32,7 +32,6 @@ use crate::{
     ffi::{AqcBidiChannel, AqcUniChannel, Ffi},
     handler::{
         BidiChannelCreated, BidiChannelReceived, Handler, UniChannelCreated, UniChannelReceived,
-        UniPsk,
     },
     shared::LabelId,
 };
@@ -412,9 +411,11 @@ pub fn test_create_bidi_channel<T: TestImpl>() {
         )
         .expect("author should be able to create a bidi channel");
 
+    let suite = CipherSuiteId::TlsAes128GcmSha256;
+
     // This is called by the author of the channel after
     // receiving the effect.
-    let author_psk = author
+    let author_secret = author
         .handler
         .bidi_channel_created(
             &mut author.eng,
@@ -430,11 +431,14 @@ pub fn test_create_bidi_channel<T: TestImpl>() {
                 psk_length_in_bytes: psk_length_in_bytes.try_into().unwrap(),
             },
         )
-        .expect("author should be able to load bidi PSK");
+        .expect("author should be able to load secret");
+    let author_psk = author_secret
+        .generate_psk(suite)
+        .expect("author should be able to generate PSK");
 
     // This is called by the channel peer after receiving the
     // effect.
-    let peer_psk = peer
+    let peer_secret = peer
         .handler
         .bidi_channel_received(
             &mut peer.eng,
@@ -450,9 +454,15 @@ pub fn test_create_bidi_channel<T: TestImpl>() {
                 psk_length_in_bytes: psk_length_in_bytes.try_into().unwrap(),
             },
         )
-        .expect("peer should be able to load bidi keys");
+        .expect("peer should be able to load secret");
+    let peer_psk = peer_secret
+        .generate_psk(suite)
+        .expect("peer should be able to generate PSK");
 
-    assert_eq!(BidiChannelId::from(channel_id), author_psk.identity());
+    assert_eq!(
+        &BidiPskId::from((channel_id.into(), suite)),
+        author_psk.identity()
+    );
     assert_eq!(author_psk.identity(), peer_psk.identity());
     assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
 }
@@ -490,9 +500,11 @@ pub fn test_create_send_only_uni_channel<T: TestImpl>() {
         )
         .expect("author should be able to create a uni channel");
 
+    let suite = CipherSuiteId::TlsAes128GcmSha256;
+
     // This is called by the author of the channel after
     // receiving the effect.
-    let author_psk = author
+    let author_secret = author
         .handler
         .uni_channel_created(
             &mut author.eng,
@@ -509,12 +521,14 @@ pub fn test_create_send_only_uni_channel<T: TestImpl>() {
                 psk_length_in_bytes: psk_length_in_bytes.try_into().unwrap(),
             },
         )
-        .expect("author should be able to load encryption key");
-    assert!(matches!(author_psk, UniPsk::SendOnly(_)));
+        .expect("author should be able to load secret");
+    let author_psk = author_secret
+        .generate_send_only_psk(suite)
+        .expect("author should be able to generate PSK");
 
     // This is called by the channel peer after receiving the
     // effect.
-    let peer_psk = peer
+    let peer_secret = peer
         .handler
         .uni_channel_received(
             &mut peer.eng,
@@ -531,10 +545,15 @@ pub fn test_create_send_only_uni_channel<T: TestImpl>() {
                 psk_length_in_bytes: psk_length_in_bytes.try_into().unwrap(),
             },
         )
-        .expect("peer should be able to load decryption key");
-    assert!(matches!(peer_psk, UniPsk::RecvOnly(_)));
+        .expect("peer should be able to load secret");
+    let peer_psk = peer_secret
+        .generate_recv_only_psk(suite)
+        .expect("peer should be able to generate PSK");
 
-    assert_eq!(UniChannelId::from(channel_id), author_psk.identity());
+    assert_eq!(
+        &UniPskId::from((channel_id.into(), suite)),
+        author_psk.identity()
+    );
     assert_eq!(author_psk.identity(), peer_psk.identity());
     assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
 }
@@ -572,9 +591,11 @@ pub fn test_create_recv_only_uni_channel<T: TestImpl>() {
         )
         .expect("author should be able to create a uni channel");
 
+    let suite = CipherSuiteId::TlsAes128GcmSha256;
+
     // This is called by the author of the channel after
     // receiving the effect.
-    let author_psk = author
+    let author_secret = author
         .handler
         .uni_channel_created(
             &mut author.eng,
@@ -591,12 +612,14 @@ pub fn test_create_recv_only_uni_channel<T: TestImpl>() {
                 psk_length_in_bytes: psk_length_in_bytes.try_into().unwrap(),
             },
         )
-        .expect("author should be able to load decryption key");
-    assert!(matches!(author_psk, UniPsk::RecvOnly(_)));
+        .expect("author should be able to load secret");
+    let author_psk = author_secret
+        .generate_recv_only_psk(suite)
+        .expect("author should be able to generate PSK");
 
     // This is called by the channel peer after receiving the
     // effect.
-    let peer_psk = peer
+    let peer_secret = peer
         .handler
         .uni_channel_received(
             &mut peer.eng,
@@ -613,10 +636,15 @@ pub fn test_create_recv_only_uni_channel<T: TestImpl>() {
                 psk_length_in_bytes: psk_length_in_bytes.try_into().unwrap(),
             },
         )
-        .expect("peer should be able to load encryption key");
-    assert!(matches!(peer_psk, UniPsk::SendOnly(_)));
+        .expect("peer should be able to load secret");
+    let peer_psk = peer_secret
+        .generate_send_only_psk(suite)
+        .expect("peer should be able to generate PSK");
 
-    assert_eq!(UniChannelId::from(channel_id), author_psk.identity());
+    assert_eq!(
+        &UniPskId::from((channel_id.into(), suite)),
+        author_psk.identity()
+    );
     assert_eq!(author_psk.identity(), peer_psk.identity());
     assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
 }
@@ -697,20 +725,29 @@ pub fn test_create_multi_bidi_channels_same_label<T: TestImpl>() {
     for (created, received) in &expect {
         // This is called by the author of the channel after
         // receiving the effect.
-        let author_psk = author
+        let author_secret = author
             .handler
             .bidi_channel_created(&mut author.eng, created)
-            .expect("author should be able to load bidi PSK");
+            .expect("author should be able to load secret");
 
         // This is called by the channel peer after receiving the
         // effect.
-        let peer_psk = peer
+        let peer_secret = peer
             .handler
             .bidi_channel_received(&mut peer.eng, received)
-            .expect("peer should be able to load bidi keys");
+            .expect("peer should be able to load secret");
 
-        assert_eq!(author_psk.identity(), peer_psk.identity());
-        assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
+        for &suite in CipherSuiteId::all() {
+            let author_psk = author_secret
+                .generate_psk(suite)
+                .expect("author should be able to generate PSK");
+            let peer_psk = peer_secret
+                .generate_psk(suite)
+                .expect("peer should be able to generate PSK");
+
+            assert_eq!(author_psk.identity(), peer_psk.identity());
+            assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
+        }
     }
 }
 
@@ -790,21 +827,30 @@ pub fn test_create_multi_bidi_channels_same_parent_cmd_id<T: TestImpl>() {
     for (created, received) in &expect {
         // This is called by the author of the channel after
         // receiving the effect.
-        let author_psk = author
+        let author_secret = author
             .handler
             .bidi_channel_created(&mut author.eng, created)
-            .expect("author should be able to load bidi PSK");
+            .expect("author should be able to load secret");
 
         // This is called by the channel peer after receiving the
         // effect.
-        let peer_psk = peer
+        let peer_secret = peer
             .handler
             .bidi_channel_received(&mut peer.eng, received)
-            .expect("peer should be able to load bidi keys");
+            .expect("peer should be able to load secret");
 
-        assert_eq!(created.channel_id, author_psk.identity());
-        assert_eq!(author_psk.identity(), peer_psk.identity());
-        assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
+        for &suite in CipherSuiteId::all() {
+            let author_psk = author_secret
+                .generate_psk(suite)
+                .expect("author should be able to generate PSK");
+            let peer_psk = peer_secret
+                .generate_psk(suite)
+                .expect("peer should be able to generate PSK");
+
+            assert_eq!(&created.channel_id, author_psk.identity().channel_id());
+            assert_eq!(author_psk.identity(), peer_psk.identity());
+            assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
+        }
     }
 }
 
@@ -894,20 +940,29 @@ pub fn test_create_multi_bidi_channels_same_label_multi_peers<T: TestImpl>() {
     for ((created, received), mut peer) in expect.iter().zip(peers) {
         // This is called by the author of the channel after
         // receiving the effect.
-        let author_psk = author
+        let author_secret = author
             .handler
             .bidi_channel_created(&mut author.eng, created)
             .expect("author should be able to load bidi PSK");
 
         // This is called by the channel peer after receiving the
         // effect.
-        let peer_psk = peer
+        let peer_secret = peer
             .handler
             .bidi_channel_received(&mut peer.eng, received)
             .expect("peer should be able to load bidi keys");
 
-        assert_eq!(author_psk.identity(), peer_psk.identity());
-        assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
+        for &suite in CipherSuiteId::all() {
+            let author_psk = author_secret
+                .generate_psk(suite)
+                .expect("author should be able to generate PSK");
+            let peer_psk = peer_secret
+                .generate_psk(suite)
+                .expect("peer should be able to generate PSK");
+
+            assert_eq!(author_psk.identity(), peer_psk.identity());
+            assert_eq!(author_psk.raw_secret_bytes(), peer_psk.raw_secret_bytes());
+        }
     }
 }