+++ crates/aranya-capi-codegen/src/ast/expand.rs
@@ -61,7 +61,7 @@ impl Ast {
             doc,
             derives,
             ext_error,
-            opaque,
+            mut opaque,
             builds,
             attrs,
             vis,
@@ -70,6 +70,9 @@ impl Ast {
             semi_token,
             ..
         } = alias;
+        if let Some(o) = &mut opaque {
+            o.generated = true;
+        }
         let strukt = Struct {
             doc,
             derives,
@@ -475,6 +478,7 @@ impl Ast {
         let err_ty = &ctx.err_ty;
 
         let doc = &f.doc;
+        let ctype_attr = parse_quote!(#[deny(improper_ctypes_definitions)]);
         let attrs = &f
             .attrs
             .iter()
@@ -482,6 +486,7 @@ impl Ast {
                 // TODO(eric): other attrs?
                 attr.path().is_ident("cfg")
             })
+            .chain(std::iter::once(&ctype_attr))
             .collect::<Vec<_>>();
 
         // Rewrite the inputs for the `extern "C"` functions and
@@ -570,7 +575,7 @@ impl Ast {
             };
 
             let pattern = format_ident!("__pattern");
-            let result = &f
+            let result = f
                 .sig
                 .output
                 .inner_type()
@@ -742,7 +747,8 @@ impl Ast {
             let block = if f_is_infallible {
                 // It's infallible, so just return the result
                 // directly.
-                quote!(#pattern)
+                let util = &ctx.util;
+                quote!(#util::check_valid_output_ty(#pattern))
             } else {
                 let success = if f.sig.output.is_result() {
                     quote! {
@@ -831,7 +837,8 @@ impl Ast {
                 let block = if f_is_infallible {
                     // It's infallible, so just return the result
                     // directly.
-                    quote!(#pattern)
+                    let util = &ctx.util;
+                    quote!(#util::check_valid_output_ty(#pattern))
                 } else {
                     // We have an output parameter, so we either
                     // return nothing or an error.