+++ crates/aranya-capi-codegen/src/syntax/node.rs
@@ -261,6 +261,7 @@ impl ToTokens for Alias {
 
         self.doc.to_tokens(tokens);
         tokens.append_all(self.attrs.outer());
+        self.opaque.to_tokens(tokens);
         self.vis.to_tokens(tokens);
         self.type_token.to_tokens(tokens);
         self.ident.to_tokens(tokens);
@@ -351,6 +352,7 @@ impl ToTokens for Struct {
         self.derives.to_tokens(tokens);
         self.repr.to_tokens(tokens);
         tokens.append_all(self.attrs.outer());
+        self.opaque.to_tokens(tokens);
         self.vis.to_tokens(tokens);
         self.struct_token.to_tokens(tokens);
         self.ident.to_tokens(tokens);
@@ -958,7 +960,7 @@ impl FfiFnSig {
                 format!("BUG: FFI functions must specify an ABI: `{}`", sig.ident),
             ));
         };
-        if !abi.name.as_ref().is_some_and(|name| name.value() == "C") {
+        if abi.name.as_ref().is_none_or(|name| name.value() != "C") {
             ctx.error(
                 &abi,
                 format!(
@@ -1345,7 +1347,7 @@ impl<T> DoubleEndedIterator for IterMut<'_, T> {
     }
 }
 
-impl<'a, T> ExactSizeIterator for IterMut<'a, T> {
+impl<T> ExactSizeIterator for IterMut<'_, T> {
     fn len(&self) -> usize {
         match self.0.as_ref() {
             Some(v) => v.len(),