+++ crates/aranya-fast-channels/src/header.rs
@@ -14,7 +14,7 @@ macro_rules! packed {
         impl $name {
             /// The size in bytes of the packed struct.
             $vis const PACKED_SIZE: usize = {
-                #[repr(packed)]
+                #[repr(C, packed)]
                 #[allow(dead_code)]
                 $vis struct $name $($tokens)*
                 ::core::mem::size_of::<$name>()